Add deelte user checking

+++ functions/APIs/todos.js
@@ -63,6 +63,9 @@ exports.deleteTodo = (request, response) => {
             if (!doc.exists) {
                 return response.status(404).json({ error: 'Todo not found' })
             }
+            if(doc.data().username !== request.user.username){
+                 return response.status(403).json({error:"UnAuthorized"})
+            }
             return document.delete();
         })
         .then(() => {
